+++ apps/frontend/src/features/openreplay/stores/openreplayStore.ts
@@ -1,13 +1,17 @@
 import { defineStore } from 'pinia'
-import { shallowRef } from 'vue'
+import { ref, shallowRef, watch } from 'vue'
 import type Tracker from '@openreplay/tracker'
 import { bus } from '@/lib/bus'
 import { useAuthStore } from '@/features/auth/stores/authStore'
 
 export const useOpenreplayStore = defineStore('openreplay', () => {
   const tracker = shallowRef<Tracker | null>(null)
+  const enabled = ref(false)
 
-  async function start(userId: string) {
+  async function start() {
+    const userId = useAuthStore().userId
+    if (!userId) return
+
     const { OPENREPLAY_PROJECT_KEY, OPENREPLAY_INGEST_POINT } = __APP_CONFIG__
     if (!OPENREPLAY_PROJECT_KEY || !OPENREPLAY_INGEST_POINT) return
 
@@ -23,23 +27,23 @@ export const useOpenreplayStore = defineStore('openreplay', () => {
     tracker.value = instance
   }
 
-  function initialize() {
-    bus.on('openreplay:start', () => {
-      const userId = useAuthStore().userId
-      if (userId) {
-        start(userId)
-      }
-    })
-  }
-
   function teardown() {
     tracker.value?.setUserID('')
     tracker.value = null
   }
 
-  return { tracker, initialize, teardown }
+  watch(enabled, (value) => {
+    if (value) {
+      start()
+    } else {
+      teardown()
+    }
+  })
+
+  return { tracker, enabled, teardown }
 })
 
 bus.on('auth:logout', () => {
-  useOpenreplayStore().teardown()
+  const store = useOpenreplayStore()
+  store.enabled = false
 })
