+++ apps/frontend/src/lib/bus.ts
@@ -4,7 +4,6 @@ import type { MessageDTO } from '@zod/messaging/messaging.dto'
 
 type AppEvents = {
   'auth:login': { token: string }
-  'openreplay:start': { userId: string | null }
   'auth:logout': void
   'auth:logged-out': void
   'auth:token-refreshed': { token: string }
